# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Meu Clube - Plataforma multiempresa para gestão de clubes e condomínios
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with phone and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Authenticated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
        "401":
          description: Invalid credentials
  /auth/password:
    patch:
      operationId: changePassword
      tags: [auth]
      summary: Change current user's password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordInput"
      responses:
        "200":
          description: Password changed
        "400":
          description: Wrong current password or invalid input
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Me"
        "401":
          description: Not authenticated

  # ── Companies ──────────────────────────────────────────────────────────────
  /companies:
    get:
      operationId: listCompanies
      tags: [companies]
      summary: List all companies (super admin)
      responses:
        "200":
          description: List of companies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Company"
    post:
      operationId: createCompany
      tags: [companies]
      summary: Create a new company
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
  /companies/{companyId}:
    get:
      operationId: getCompany
      tags: [companies]
      summary: Get company by ID
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Company
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
    patch:
      operationId: updateCompany
      tags: [companies]
      summary: Update company
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
  /companies/{companyId}/status:
    patch:
      operationId: updateCompanyStatus
      tags: [companies]
      summary: Change company status (activate, suspend, cancel)
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyStatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"

  # ── Company Users ──────────────────────────────────────────────────────────
  /companies/{companyId}/users:
    get:
      operationId: listCompanyUsers
      tags: [users]
      summary: List users linked to a company
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of company users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CompanyUser"
    post:
      operationId: addCompanyUser
      tags: [users]
      summary: Add or invite a user to a company
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyUserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanyUser"
...
  /companies/{companyId}/responsible:
    patch:
      operationId: setCompanyResponsible
      tags: [companies]
      summary: Set the responsible user for a company
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetResponsibleInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
  /companies/{companyId}/users/{userId}:
    patch:
      operationId: updateCompanyUser
      tags: [users]
      summary: Update user role or status in a company
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyUserUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanyUser"
    delete:
      operationId: removeCompanyUser
      tags: [users]
      summary: Remove user from a company
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed

  # ── Structures ─────────────────────────────────────────────────────────────
...
  /users:
    get:
      operationId: listAdminUsers
      tags: [users]
      summary: List all super admin users
      responses:
        "200":
          description: Super admin users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminUser"
    post:
      operationId: createAdminUser
      tags: [users]
      summary: Create a new super admin user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAdminUserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "409":
          description: Phone already in use
  /users/{userId}:
    delete:
      operationId: deleteAdminUser
      tags: [users]
      summary: Delete a super admin user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Plans ──────────────────────────────────────────────────────────────────
...
components:
  schemas:
    AdminUser:
      type: object
      required: [id, name, phone, role, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: string
          nullable: true
        role:
          type: string
        createdAt:
          type: string
    CreateAdminUserInput:
      type: object
      required: [name, phone, password]
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        password:
          type: string
          minLength: 6
...
    LoginInput:
      type: object
      required: [phone, password]
      properties:
        phone:
          type: string
        password:
          type: string
    AuthSession:
      type: object
      required: [id, name, phone, role]
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        role:
          type: string
          description: "Global role: super_admin or user"
        companyId:
          type: ["integer", "null"]
          description: Active company context
    Me:
      type: object
      required: [id, name, phone, role]
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        role:
          type: string
        companyId:
          type: ["integer", "null"]
        companyRole:
          type: ["string", "null"]
          description: "Role within active company: admin, operator or member"

    # ── Companies ────────────────────────────────────────────────────────────
    Company:
      type: object
      required: [id, name, type, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          description: "club, condominium or other"
        status:
          type: string
          description: "active, inactive, suspended, cancelled"
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        planId:
          type: ["integer", "null"]
        responsibleUserId:
          type: ["integer", "null"]
          description: "ID of the responsible user for this company"
    ChangePasswordInput:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
...
    SetResponsibleInput:
      type: object
      required: [userId]
      properties:
        userId:
          type: integer
          description: "ID of the user to set as responsible"
    CompanyInput:
      type: object
      required: [name, type]
      properties:
        name:
          type: string
        type:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        adminName:
          type: string
          description: "Name of the initial admin user"
        adminPhone:
          type: string
          description: "Phone (login) of the initial admin user"
        adminPassword:
          type: string
          description: "Password for the initial admin user"
    CompanyUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
    CompanyStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          description: "active, inactive, suspended, cancelled"

    # ── Users ────────────────────────────────────────────────────────────────
    CompanyUser:
      type: object
      required: [id, userId, companyId, role, status, name, phone]
      properties:
        id:
          type: integer
        userId:
          type: integer
        companyId:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        role:
          type: string
          description: "admin, operator, member"
        status:
          type: string
          description: "active, inactive"
        createdAt:
          type: string
          format: date-time
    CompanyUserInput:
      type: object
      required: [name, phone, role]
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
          description: "admin, operator, member"
    CompanyUserUpdate:
      type: object
      properties:
        role:
          type: string
        status:
          type: string

    # ── Structures ────────────────────────────────────────────────────────────
...