# lib/api-zod/src/generated/types/authSession.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Meu Clube - Plataforma multiempresa para gestão de clubes e condomínios
 * OpenAPI spec version: 0.1.0
 */

export interface AuthSession {
  id: number;
  name: string;
  phone: string;
  /** Global role: super_admin or user */
  role: string;
  /**
     * Active company context
     * @nullable
     */
  companyId?: number | null;
}