# lib/api-zod/src/generated/types/company.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Meu Clube - Plataforma multiempresa para gestão de clubes e condomínios
 * OpenAPI spec version: 0.1.0
 */

export interface Company {
  id: number;
  name: string;
  /** club, condominium or other */
  type: string;
  /** active, inactive, suspended, cancelled */
  status: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  address?: string | null;
  createdAt: Date;
  /** @nullable */
  planId?: number | null;
  /**
     * ID of the responsible user for this company
     * @nullable
     */
  responsibleUserId?: number | null;
}