# lib/api-zod/src/generated/types/createAdminUserInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Meu Clube - Plataforma multiempresa para gestão de clubes e condomínios
 * OpenAPI spec version: 0.1.0
 */

export interface CreateAdminUserInput {
  name: string;
  phone: string;
  email?: string;
  /** @minLength 6 */
  password: string;
}